requirements.txt
google-generativeai==0.3.2
streamlit==1.31.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
plotly==5.18.0
pandas==2.2.0
numpy==1.26.4

# # Fixed Requirements - Compatible Versions

# # Core OpenAI & LangChain
# openai==1.12.0
# langchain==0.1.10
# langchain-community==0.0.25
# langchain-openai==0.0.6

# # Vector Database
# faiss-cpu==1.8.0

# # Web Frameworks
# streamlit==1.31.0
# fastapi==0.109.2
# uvicorn[standard]==0.27.1

# # Utilities
# python-dotenv==1.0.1
# pydantic==2.6.1
# httpx==0.26.0

# # Data Science
# numpy==1.26.4
# pandas==2.2.0
# plotly==5.18.0
# scikit-learn==1.4.0

# # Additional dependencies
# tiktoken==0.5.2